refactor common V2/V3 code into ParserBase
