+++ lib/Parser.v2.js
@@ -1,5 +1,5 @@
 // a class for parsing openapi V2 data into config data for fastify
-import { ParserBase, HttpOperations } from "./ParserBase.js";
+import { ParserBase } from "./ParserBase.js";
 
 export class ParserV2 extends ParserBase {
 	parseParams(data) {
@@ -108,50 +108,6 @@ export class ParserV2 extends ParserBase {
 		// remove loops from the schema so fastify wont break
 		this.removeRecursion(schema);
 		return schema;
-	}
-
-	processOperation(path, operation, operationSpec, genericSchema) {
-		const route = {
-			method: operation.toUpperCase(),
-			url: this.makeURL(path),
-			schema: this.makeSchema(genericSchema, operationSpec),
-			openapiPath: path,
-			operationId:
-				operationSpec.operationId || this.makeOperationId(operation, path),
-			openapiSource: operationSpec,
-			security: this.parseSecurity(
-				operationSpec.security || this.spec.security,
-			),
-		};
-
-		if (operationSpec["x-fastify-config"]) {
-			route.config = operationSpec["x-fastify-config"];
-		}
-
-		this.config.routes.push(route);
-	}
-
-	processPaths(paths) {
-		const copyItems = ["summary", "description"];
-		for (const path in paths) {
-			const genericSchema = {};
-			const pathSpec = paths[path];
-
-			this.copyProps(pathSpec, genericSchema, copyItems, true);
-			if (typeof pathSpec.parameters === "object") {
-				this.parseParameters(genericSchema, pathSpec.parameters);
-			}
-			for (const operation in pathSpec) {
-				if (HttpOperations.has(operation)) {
-					this.processOperation(
-						path,
-						operation,
-						pathSpec[operation],
-						genericSchema,
-					);
-				}
-			}
-		}
 	}
 
 	parse(spec) {
