+++ lib/Parser.v3.js
@@ -1,6 +1,6 @@
 // a class for parsing openapi v3 data into config data for fastify
 
-import { ParserBase, HttpOperations } from "./ParserBase.js";
+import { ParserBase } from "./ParserBase.js";
 
 function isExploding(item) {
 	const explode = item.explode ?? item.style === "form" ? true : false;
@@ -121,50 +121,6 @@ export class ParserV3 extends ParserBase {
 
 		this.removeRecursion(schema);
 		return schema;
-	}
-
-	processOperation(path, operation, operationSpec, genericSchema) {
-		const route = {
-			method: operation.toUpperCase(),
-			url: this.makeURL(path),
-			schema: this.makeSchema(genericSchema, operationSpec),
-			openapiPath: path,
-			operationId:
-				operationSpec.operationId || this.makeOperationId(operation, path),
-			openapiSource: operationSpec,
-			security: this.parseSecurity(
-				operationSpec.security || this.spec.security,
-			),
-		};
-
-		if (operationSpec["x-fastify-config"]) {
-			route.config = operationSpec["x-fastify-config"];
-		}
-
-		this.config.routes.push(route);
-	}
-
-	processPaths(paths) {
-		const copyItems = ["summary", "description"];
-		for (const path in paths) {
-			const genericSchema = {};
-			const pathSpec = paths[path];
-
-			this.copyProps(pathSpec, genericSchema, copyItems, true);
-			if (typeof pathSpec.parameters === "object") {
-				this.parseParameters(genericSchema, pathSpec.parameters);
-			}
-			for (const operation in pathSpec) {
-				if (HttpOperations.has(operation)) {
-					this.processOperation(
-						path,
-						operation,
-						pathSpec[operation],
-						genericSchema,
-					);
-				}
-			}
-		}
 	}
 
 	parse(spec) {
