+++ lib/ParserBase.js
@@ -1,5 +1,15 @@
 // baseclass for the v2 and v3 parsers
 
+export const HttpOperations = new Set([
+	"delete",
+	"get",
+	"head",
+	"patch",
+	"post",
+	"put",
+	"options",
+]);
+
 export class ParserBase {
 	constructor() {
 		this.config = { generic: {}, routes: [], contentTypes: new Set() };
@@ -98,14 +108,48 @@ export class ParserBase {
 			}
 		}
 	}
-}
 
-export const HttpOperations = new Set([
-	"delete",
-	"get",
-	"head",
-	"patch",
-	"post",
-	"put",
-	"options",
-]);
+	processOperation(path, operation, operationSpec, genericSchema) {
+		const route = {
+			method: operation.toUpperCase(),
+			url: this.makeURL(path),
+			schema: this.makeSchema(genericSchema, operationSpec),
+			openapiPath: path,
+			operationId:
+				operationSpec.operationId || this.makeOperationId(operation, path),
+			openapiSource: operationSpec,
+			security: this.parseSecurity(
+				operationSpec.security || this.spec.security,
+			),
+		};
+
+		if (operationSpec["x-fastify-config"]) {
+			route.config = operationSpec["x-fastify-config"];
+		}
+
+		this.config.routes.push(route);
+	}
+
+	processPaths(paths) {
+		const copyItems = ["summary", "description"];
+		for (const path in paths) {
+			const genericSchema = {};
+			const pathSpec = paths[path];
+
+			this.copyProps(pathSpec, genericSchema, copyItems, true);
+			if (typeof pathSpec.parameters === "object") {
+				this.parseParameters(genericSchema, pathSpec.parameters);
+			}
+			for (const operation in pathSpec) {
+				if (HttpOperations.has(operation)) {
+					this.processOperation(
+						path,
+						operation,
+						pathSpec[operation],
+						genericSchema,
+					);
+				}
+			}
+		}
+	}
+}
